+++ package.json
@@ -2,7 +2,7 @@
   "name": "sunflower-land-timer",
   "version": "1.0.0",
   "scripts": {
-    "start": "expo start --dev-client",
+    "start": "expo start --tunnel",
     "dev": "expo start --dev-client",
     "android": "expo run:android",
     "ios": "expo run:ios",
@@ -12,7 +12,6 @@
     "@expo/prebuild-config": "5.0.7",
     "@expo/vector-icons": "^13.0.0",
     "@expo/webpack-config": "^0.17.2",
-    "@notifee/react-native": "^7.3.0",
     "@react-native-async-storage/async-storage": "~1.17.3",
     "@react-navigation/bottom-tabs": "^6.5.2",
     "@react-navigation/native": "^6.1.1",
@@ -21,14 +20,17 @@
     "expo": "^47.0.13",
     "expo-background-fetch": "^11.0.1",
     "expo-build-properties": "^0.4.1",
+    "expo-constants": "^14.0.2",
     "expo-dev-client": "^2.0.1",
+    "expo-device": "^5.0.0",
+    "expo-notifications": "^0.17.0",
     "expo-splash-screen": "^0.17.5",
     "expo-status-bar": "~1.4.2",
     "expo-task-manager": "^11.0.1",
     "expo-updates": "~0.15.6",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-native": "^0.70.6",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-native": "0.70.5",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.2",
     "react-native-vector-icons": "^9.2.0",
