+++ src/TimerView/List.jsx
@@ -1,9 +1,8 @@
 import { StyleSheet, Text, View, Button, Image, TouchableOpacity } from 'react-native';
 import { useState, useContext, useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import getStore from '../../utils/getStore';
 import stateTrue from '../../utils/stateTrue';
-import stateEnd from '../../utils/stateEnd';
 import Timer from './Timer';
 
 export default function ItemList({ item }) {
@@ -14,16 +13,13 @@ export default function ItemList({ item }) {
   const startTimer = () => {
     stateTrue({ dispatch, itemName });
   };
-  const endTimer = () => {
-    stateEnd({ dispatch, itemName });
-  };
 
   return (
     <TouchableOpacity stle={styles.startBtn} onPress={() => startTimer()}>
       <View style={[styles.itemContainer, storeData.isActive ? styles.active : styles.notActive]}>
         <Image style={styles.itemLogo} source={item.image} />
         <View style={styles.textContainer}>
-          <Timer itemName={itemName} data={storeData} isActive={storeData.isActive} endTimer={() => endTimer()} key={itemName} dispatch={dispatch} />
+          <Timer itemName={itemName} data={storeData} isActive={storeData.isActive} key={itemName} dispatch={dispatch} />
         </View>
         {storeData.isActive ? null : <Text style={{ color: '#742C2C' }}>Press to Start</Text>}
       </View>
